댓글기능 구현

+++ auctionBay/src/main/java/com/kh/auctionBay/board/controller/CommentApiController.java
@@ -34,17 +34,17 @@ public class CommentApiController {
 															  @RequestBody CommentRequest commentRequest,
 															  HttpSession session) {
 		
-		
 		// 로그인한 사용자 정보 추출
-		UserDTO loginMember = (UserDTO)session.getAttribute(SessionConst.LOGIN_USER);
-		if (loginMember == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                                 .body(ApiResponse.fail("로그인이 필요한 서비스입니다."));
-        }
+		UserDTO loginUser = (UserDTO)session.getAttribute(SessionConst.LOGIN_USER);
+		if (loginUser == null) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+					.body(ApiResponse.fail("로그인이 필요한 서비스입니다."));
+		}
+		
 		// 게시글 번호, 사용자 아이디, 댓글 정보를 저장
 		try {
 			
-			CommentDTO comment = service.addComment(productId, commentRequest.getContent(), loginMember.getUserNo());
+			CommentDTO comment = service.addComment(productId, commentRequest.getContent(), loginUser.getUserNo());
 
 			return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.success(comment));
 		
@@ -61,8 +61,12 @@ public class CommentApiController {
 	@DeleteMapping("/comments/{commentId}")
 	public ResponseEntity<ApiResponse<Long>> deleteComment(@PathVariable Long commentId,
 														   HttpSession session) {
-		
+	 
 		UserDTO loginUser = (UserDTO) session.getAttribute(SessionConst.LOGIN_USER);
+		if (loginUser == null) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+					.body(ApiResponse.fail("로그인이 필요한 서비스입니다."));
+		}
 		
 		try {
 			service.deleteComment(commentId, loginUser.getUserNo());
